refactor(symbol_analyzer): split into mod/markdown/tests submodule
`symbol_analyzer.rs` was 1609 LOC and tripped the "Reject .rs files
over 1000 LOC" CI gate. Restructure into a directory module with
three files, each well under the limit:

| File | LOC | Contents |
|------|-----|----------|
| `mod.rs` | 608 | ELF probe, `analyze_elf` + `AnalyzeConfig`, nm/c++filt drivers, `format_text_report`, project ELF discovery |
| `markdown.rs` | 430 | `format_markdown_report`, `format_markdown_report_with_graphs`, `MarkdownGraphOptions`, `SidecarOptions`, `write_sidecar_dot_files`, internal `emit_*` helpers |
| `tests.rs` | 597 | All 17 `#[cfg(test)]` cases |

The markdown public surface is re-exported from `mod.rs`
(`pub use markdown::{format_markdown_report, ...}`) so existing
consumers like `fbuild_cli::cli::symbols_cmd` and
`fbuild_cli::cli::graph_cmd` keep importing
`fbuild_build::symbol_analyzer::format_markdown_report` unchanged.

Content-preserving — no behavior change. The only non-mechanical edit
is two missing local imports in one test
(`markdown_report_emits_dual_ranked_callees_subtable`) that
previously resolved through the file's module-scope imports and now
need explicit `use super::*;` + `use super::markdown::*;` after the
split.

Verified clean: `check --workspace --all-targets`,
`clippy --workspace --all-targets -- -D warnings`,
`test -p fbuild-build --lib symbol_analyzer` (17/17),
`fmt --check`.
